refactor: useS3Upload 반환 타입 변경
